+++ src/characteristic.rs
@@ -151,13 +151,13 @@ impl Characteristic {
             char.set_write_type(env, write_type)?;
             if android_api_level() >= 33 {
                 conn.gatt
-                    .as_new_api()
+                    .as_new_api(env)
                     .write_characteristic(env, char, array, write_type)?
                     .check_status_code()
             } else {
-                char.as_old_api().set_value(env, array)?;
+                char.as_old_api(env).set_value(env, array)?;
                 conn.gatt
-                    .as_old_api()
+                    .as_old_api(env)
                     .write_characteristic(env, char)
                     .map_err(|e| e.into())
                     .and_then(|b| b.non_false())